moving handleChangelogEntriesParsing, handleManualChangesetCreation and   handleAutomaticChangesetCreation util functions to handlers.utils.js
